pyproject.toml
[tool.poetry]
name = "ICU"
version = "0.1.0"
description = ""
authors = ["Bap_Le_Goff <baptiste_legoff@outlook.com>"]

[tool.poetry.dependencies]
python = "^3.8"
matplotlib = "3.2.2"
numpy = "1.18.5"
Pillow = "7.1.2"
PyYAML = "5.3.1"
requests = "2.23.0"
scipy = "1.4.1"
torch = "1.7.0"
torchvision = "0.8.1"
tqdm = "4.41.0"
tensorboard = "2.4.1"
pandas = "1.1.4"
seaborn = "0.11.0"
streamlit = "1.4"
mysql-connector-python = "^8.0.28"

[tool.poetry.dev-dependencies]
pytest = "^5.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
